stage,test: extract make_container() test helper
