+++ osbuild/testutil/__init__.py
@@ -6,6 +6,7 @@ import os
 import pathlib
 import re
 import shutil
+import subprocess
 import tempfile
 
 
@@ -77,3 +78,20 @@ def mock_command(cmd_name: str, script: str):
             yield
         finally:
             os.environ["PATH"] = original_path
+
+
+def make_container(tmp_path, tag, fake_content, base="scratch"):
+    fake_container_src = tmp_path / "fake-container-src"
+    make_fake_tree(fake_container_src, fake_content)
+    fake_containerfile_path = fake_container_src / "Containerfile"
+    container_file_content = f"""
+    FROM {base}
+    COPY . .
+    """
+    fake_containerfile_path.write_text(container_file_content, encoding="utf8")
+    subprocess.check_call([
+        "podman", "build",
+        "--no-cache",
+        "-f", os.fspath(fake_containerfile_path),
+        "-t", tag,
+    ])
